backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
PyJWT==2.8.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
flask-caching==2.1.0
